+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 ### v1.0.2
 * Updated all my mods to use a single helper file.
 * Updated installation instructions to use the mod manager instead.
+
+### v1.1.0
+* Now synced in online multiplayer.
+* Rewards are now spawned as command crates so that they take up less room.
+    * Now spawns one 3-item crate per player.
+* Gamemode has been rebalanced overall.
+    * Minimum waves for completion raised to 20.
+    * Infusion and Umbrella can no longer spawn.